Adding auto refresh for the S3 storage provider
Adding a new component, watcher, that can be used to implement
storage watchers. May move the local watcher into the new package
at some point, but it is not necessary right now.

Changed behavior in the main program to not stop the watcher if
the watcher encounters an error. Errors could be transitory, so
it does not make sense to close the watcher immediately on one
error.

Changed the name of the CloseLogger() to CloseLogging() to make it more
clear.

Also commented the code a bit more.