Large refactor to simplify the codebase
This was a large refactor to simplify the codebase. There were functions
that did similiar things in diferent files, so those were combined. The
refactor also made signing key password management more consistent and
fixed an issue where signing providers other than local can be used for
signing packages.

There is still some work to be done in this area, but this commit lays
the groundwork.