Signing provider and command line updates
Updated the command line options to allow for all commands to have
signing and storage providers specified.

Added two options to the local signing provider: key and copy. These
will allow for signing keys to be used and loaded from alternate
locations.

Fixed a small issue in the Vault singing provider where the wrong
internal name for the provider was being used.

Enforced the need for a signing provider to be provided when signing an
index or package using the sign commands when an external signing
provider is specified in the provided config.